NewsList rendering managed in app class
- app class has loadingNewsList state
- app class renders NewsList class
- Header home button onclick alters app class state thus rendering news list
- removed newslist rendering from dashboard
