+++ src/components/Dashboard.js
@@ -3,7 +3,6 @@ import PortfolioList from "./Portfolio/PortfolioList";
 import TransactionList from "./Portfolio/Transactions/TransactionList";
 import Header from "./Header";
 import axios from 'axios';
-import NewsList from './newsfeed/NewsList'
 import PortfolioForm from './Portfolio/PortfolioForm'
 
 
@@ -13,14 +12,12 @@ export default class Dashboard extends Component {
     this.state = {
       portfolioID: null,
       portfolioName: '',
-      loadingNewsList: false,
       loadingPortfolioForm: false
     }
 
     this.loadPortfolio = this.loadPortfolio.bind(this)
     this.loadDashboard = this.loadDashboard.bind(this)
     this.deletePortfolio = this.deletePortfolio.bind(this)
-    this.loadNewsList = this.loadNewsList.bind(this)
     this.loadPortfolioForm = this.loadPortfolioForm.bind(this)
   }
 
@@ -37,20 +34,12 @@ export default class Dashboard extends Component {
     })
   }
 
-  loadNewsList(){
-    this.setState({
-      loadingNewsList: true,
-      portfolioID: null,
-      loadingPortfolioForm: false
-    })
-  }
-
   loadDashboard(){
     this.setState({
       portfolioID: null,
-      loadingNewsList: false,
       loadingPortfolioForm: false
     })
+    this.props.loadHomePage()
   }
 
   deletePortfolio(event){
@@ -67,16 +56,15 @@ export default class Dashboard extends Component {
 
   render(){
     let noPortfolioId = (this.state.portfolioID === null)
-    let loadingNewsList = (this.state.loadingNewsList === true)
     let loadingPortfolioForm = (this.state.loadingPortfolioForm === true)
-    if (noPortfolioId && !loadingNewsList && !loadingPortfolioForm) {
+    if (noPortfolioId && !loadingPortfolioForm) {
       return (
         <div class="dashboard_background2">
           <Header 
           userToken={this.props.userToken} 
           handleLogout={this.props.handleLogout}
           loadDashboard={this.loadDashboard}
-          loadNewsList={this.loadNewsList}
+          loadNewsList={this.props.loadNewsList}
           />
           <div class="div1">
             <h1 classname="mt-5">My Portfolios</h1>
@@ -85,28 +73,14 @@ export default class Dashboard extends Component {
           </div>
         </div>
       )
-    } else if (noPortfolioId && loadingNewsList && !loadingPortfolioForm){
-      return(
-        <div class="dashboard_background2">
-          <Header 
-            userToken={this.props.userToken} 
-            handleLogout={this.props.handleLogout}
-            loadDashboard={this.loadDashboard}
-            loadNewsList={this.loadNewsList}
-          />
-          <div class="div1">
-            <NewsList />
-          </div>
-        </div>
-      )
-    } else if (noPortfolioId && !loadingNewsList && loadingPortfolioForm) {
+    } else if (noPortfolioId && loadingPortfolioForm) {
       return (
         <div class="dashboard_background2">
           <Header 
             userToken={this.props.userToken} 
             handleLogout={this.props.handleLogout}
             loadDashboard={this.loadDashboard}
-            loadNewsList={this.loadNewsList}
+            loadNewsList={this.props.loadNewsList}
           />
         <div class="container">
             <PortfolioForm loadDashboard={this.loadDashboard} userToken={this.props.userToken}/> 
@@ -120,7 +94,7 @@ export default class Dashboard extends Component {
               userToken={this.props.userToken} 
               handleLogout={this.props.handleLogout}
               loadDashboard={this.loadDashboard}
-              loadNewsList={this.loadNewsList}
+              loadNewsList={this.props.loadNewsList}
             />
             <div class="div1">
               <TransactionList
